+++ sync-for-payables/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v6.0.0] sync-for-payables
 ### Releases
 - [NPM v6.0.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/6.0.0 - sync-for-payables
+
+## 2024-10-30 11:20:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.423.2 (2.442.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v7.0.0] sync-for-payables
+### Releases
+- [NPM v7.0.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/7.0.0 - sync-for-payables